+++ core/src/main/java/de/muenchen/rbs/kitafinderdatenservice/repository/BewerbungRepository.java
@@ -11,6 +11,6 @@ public interface BewerbungRepository
 		extends PagingAndSortingRepository<Bewerbung, ExportId>, CrudRepository<Bewerbung, ExportId> {
 
 	@Modifying
-	int deleteByExportId(Integer exportId);
+	int deleteByExportId(Long exportId);
 
 }
