+++ core/src/main/java/de/muenchen/rbs/kitafinderdatenservice/repository/ExportRunRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import de.muenchen.rbs.kitafinderdatenservice.domain.ExportRun;
 
-public interface ExportRunRepository extends CrudRepository<ExportRun, Integer> {
+public interface ExportRunRepository extends CrudRepository<ExportRun, Long> {
 
 	@Query(value = "SELECT e FROM ExportRun e WHERE e.status = de.muenchen.rbs.kitafinderdatenservice.domain.ExportStatus.SUCCESS ORDER BY e.startTime DESC")
 	List<ExportRun> findAllSuccessfullOrdered();
